Compose strategies to decoding_strategy
