+++ dev_tool/show_encoding.py
@@ -9,6 +9,7 @@ vint_root = Path(__file__).resolve().parent.parent
 sys.path.insert(0, str(vint_root))
 
 from vint.encodings.decoder import Decoder
+from vint.encodings.decoding_strategy import default_decoding_strategy
 
 
 if __name__ == '__main__':
@@ -17,6 +18,6 @@ if __name__ == '__main__':
     namespace = vars(arg_parser.parse_args(sys.argv[1:]))
 
     file_path = Path(namespace['file'][0])
-    decoder = Decoder(Decoder.default_strategy())
+    decoder = Decoder(default_decoding_strategy)
     decoder.read(file_path)
     pprint(decoder.debug_hint)
